refactor(opening hours): move to own file
- moved opening hours helper to own file
- added tests for isOpen
  - added new arg to isOpen for testing purposes

SVA-319
