+++ src/helpers/index.js
@@ -16,6 +16,7 @@ export * from './linkHelper';
 export * from './mapHelper';
 export * from './matomoHelper';
 export * from './momentHelper';
+export * from './openingHoursHelper';
 export * from './personNameHelper';
 export * from './positionHelper';
 export * from './priceHelper';
